rename user (controller package) to priv
